# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API для платформы по усыновлению домашних животных.",
        "title": "Happytail API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/shelters": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shelters"
                ],
                "summary": "Получить все приюты",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Shelter"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shelters"
                ],
                "summary": "Создать приют",
                "parameters": [
                    {
                        "description": "данные приюта",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.CreateShelterInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Shelter"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shelters/nearby": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shelters"
                ],
                "summary": "Найти ближайшие приюты",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Широта",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Долгота",
                        "name": "lon",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Радиус поиска в км (default: 10, max: 500)",
                        "name": "radius",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.ShelterWithDistance"
                            }
                        }
                    }
                }
            }
        },
        "/shelters/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shelters"
                ],
                "summary": "Получить приют по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID приюта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Shelter"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет приют по ID (soft delete)",
                "tags": [
                    "shelters"
                ],
                "summary": "Удалить приют",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID приюта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно удалено"
                    },
                    "400": {
                        "description": "Invalid shelter ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Shelter not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to delete shelter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shelters"
                ],
                "summary": "Обновить данные приюта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID приюта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "поля для обновления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.UpdateShelterInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Shelter"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shelters/{id}/animals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shelters"
                ],
                "summary": "Получить животных приюта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID приюта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Страница",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по типу",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.PaginatedResponse-github_com_sd0hni-psina_happytail_internal_models_Animal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sd0hni-psina_happytail_internal_models.Animal": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "breed": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_vet_passport": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_vaccinated": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "share_count": {
                    "type": "integer"
                },
                "shelter_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.AnimalType"
                }
            }
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.AnimalType": {
            "type": "string",
            "enum": [
                "dog",
                "cat",
                "bunny"
            ],
            "x-enum-varnames": [
                "AnimalTypeDog",
                "AnimalTypeCat",
                "AnimalTypeBunny"
            ]
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.CreateShelterInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.PaginatedResponse-github_com_sd0hni-psina_happytail_internal_models_Animal": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Animal"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.PaginationMeta"
                }
            }
        },
        "github_com_sd0hni-psina_happytail_internal_models.PaginationMeta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.Shelter": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "github_com_sd0hni-psina_happytail_internal_models.ShelterWithDistance": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "distance_km": {
                    "type": "number"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.UpdateShelterInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}